requirements.txt
httpx
aiohttp
python-dotenv 
google-generativeai
langgraph
langchain
pydantic
tavily-python
bs4
requests
pdfminer.six
